Basic DegreeRequirements working. Added is_admin to User. Refactored _errors partial, AuthId middleware and admin middleware

+++ resources/views/student/index.blade.php
@@ -11,7 +11,9 @@
                 <th>NetID</th>
                 <th>NUID</th>
                 <th>Status</th>
-                <th>Operations</th>
+                @if(Auth::user()->is_admin)
+                    <th>Operations</th>
+                @endif
             </tr>
         </thead>
         <tbody>
@@ -22,13 +24,15 @@
                     <td>{{ $student->netid }}</td>
                     <td>{{ $student->nuid }}</td>
                     <td>{{ $student->status }}</td>
-                    <td>
-                        {!! Form::model($student, ['method'=>'delete', 'class'=>'delete_confirm',
-                                   'action'=>['StudentController@destroy', $student->id]]) !!}
-                            <a href="{{ action('StudentController@edit', $student->id) }}" class="btn btn-cta-red"><i class="fa fa-edit"></i> Edit</a>
-                            <button type="submit" class="btn"><i class="fa fa-trash"></i> Delete</button>
-                        {!! Form::close() !!}
-                    </td>
+                    @if(Auth::user()->is_admin)
+                        <td>
+                            {!! Form::model($student, ['method'=>'delete', 'class'=>'delete_confirm',
+                                       'action'=>['StudentController@destroy', $student->id]]) !!}
+                                <a href="{{ action('StudentController@edit', $student->id) }}" class="btn btn-cta-red"><i class="fa fa-edit"></i> Edit</a>
+                                <button type="submit" class="btn"><i class="fa fa-trash"></i> Delete</button>
+                            {!! Form::close() !!}
+                        </td>
+                    @endif
                 </tr>
             @endforeach
         </tbody>
